requirements.txt
django==3.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django_filter==21.1